server/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==2.5.0
python-dotenv==1.0.0
pydantic-settings==2.1.0
# Vertex AI dependencies
google-cloud-aiplatform==1.38.1
vertexai==1.38.1
google-auth==2.23.4
numpy==1.24.3
google-cloud-aiplatform==1.38.1
google-auth==2.24.0